Update use of baths to be BigDecimal

+++ backend/src/main/java/com/propertystack/homematch/Listing/ListingFilter.java
@@ -6,7 +6,7 @@ public record ListingFilter (
     BigDecimal minPrice,
     BigDecimal maxPrice,
     Integer minBeds,
-    Double minBaths,
+    BigDecimal  minBaths,
     Integer minSqft
 ) {
